# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Audio Player API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Track {
  id: number;
  title: string;
  artist: string;
  album: string;
  /** @nullable */
  year?: number | null;
  /** @nullable */
  genre?: string | null;
  duration: number;
  /** @nullable */
  trackNumber?: number | null;
  fileName: string;
  folderPath: string;
  /** @nullable */
  albumArtDataUrl?: string | null;
  rating: number;
  source: string;
  /** @nullable */
  subsonicId?: string | null;
  /** @nullable */
  subsonicServerId?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateTrackBody {
  title: string;
  artist?: string;
  album?: string;
  /** @nullable */
  year?: number | null;
  /** @nullable */
  genre?: string | null;
  duration?: number;
  /** @nullable */
  trackNumber?: number | null;
  fileName: string;
  folderPath?: string;
  /** @nullable */
  albumArtDataUrl?: string | null;
  rating?: number;
  source?: string;
  /** @nullable */
  subsonicId?: string | null;
  /** @nullable */
  subsonicServerId?: number | null;
}

export interface UpdateTrackBody {
  title?: string;
  artist?: string;
  album?: string;
  /** @nullable */
  year?: number | null;
  /** @nullable */
  genre?: string | null;
  /** @nullable */
  trackNumber?: number | null;
  rating?: number;
  /** @nullable */
  albumArtDataUrl?: string | null;
}

export interface BulkUpsertTracksBody {
  tracks: CreateTrackBody[];
}

export interface BulkUpsertTracksResponse {
  upserted: number;
  tracks: Track[];
}

export interface Playlist {
  id: number;
  name: string;
  /** @nullable */
  query?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreatePlaylistBody {
  name: string;
  /** @nullable */
  query?: string | null;
}

export interface UpdatePlaylistBody {
  name?: string;
  /** @nullable */
  query?: string | null;
}

export interface PlaylistTrack {
  id: number;
  playlistId: number;
  trackId: number;
  position: number;
}

export interface AddTrackToPlaylistBody {
  trackId: number;
  position?: number;
}

export interface QueueItem {
  id: number;
  trackId: number;
  position: number;
  track: Track;
}

export interface ReplaceQueueBody {
  trackIds: number[];
}

export interface AddToQueueBody {
  trackIds: number[];
}

export interface EqPreset {
  id: number;
  name: string;
  bands: string;
  isActive: boolean;
  isBuiltin: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreateEqPresetBody {
  name: string;
  bands: string;
  isActive?: boolean;
}

export interface UpdateEqPresetBody {
  name?: string;
  bands?: string;
}

export interface SetActiveEqPresetBody {
  /** @nullable */
  presetId: number | null;
}

export interface SubsonicServer {
  id: number;
  name: string;
  url: string;
  username: string;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreateSubsonicServerBody {
  name: string;
  url: string;
  username: string;
  password: string;
  isActive?: boolean;
}

export interface UpdateSubsonicServerBody {
  name?: string;
  url?: string;
  username?: string;
  password?: string;
  isActive?: boolean;
}

export interface SubsonicTestResult {
  success: boolean;
  message: string;
  /** @nullable */
  version?: string | null;
}

export type ListTracksParams = {
  search?: string;
  artist?: string;
  album?: string;
  genre?: string;
  sortBy?: string;
  sortDir?: string;
};