# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Audio Player API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /playlists:
    get:
      operationId: listPlaylists
      tags: [playlists]
      summary: List all playlists
      responses:
        "200":
          description: Playlists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Playlist"
    post:
      operationId: createPlaylist
      tags: [playlists]
      summary: Create a playlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlaylistBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Playlist"
  /playlists/{id}:
    get:
      operationId: getPlaylist
      tags: [playlists]
      summary: Get a playlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Playlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Playlist"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePlaylist
      tags: [playlists]
      summary: Update a playlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlaylistBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Playlist"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deletePlaylist
      tags: [playlists]
      summary: Delete a playlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /playlists/{id}/tracks:
    get:
      operationId: getPlaylistTracks
      tags: [playlists]
      summary: Get tracks for a playlist (evaluates smart playlist query if present)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tracks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Track"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: addTrackToPlaylist
      tags: [playlists]
      summary: Add a track to a playlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddTrackToPlaylistBody"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaylistTrack"
  /playlists/{id}/tracks/{trackId}:
    delete:
      operationId: removeTrackFromPlaylist
      tags: [playlists]
      summary: Remove a track from a playlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: trackId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
  /queue:
    get:
      operationId: getQueue
      tags: [queue]
      summary: Get current play queue
      responses:
        "200":
          description: Queue
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QueueItem"
    put:
      operationId: replaceQueue
      tags: [queue]
      summary: Replace entire queue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplaceQueueBody"
      responses:
        "200":
          description: Updated queue
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QueueItem"
    post:
      operationId: addToQueue
      tags: [queue]
      summary: Add track(s) to queue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToQueueBody"
      responses:
        "200":
          description: Updated queue
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QueueItem"
    delete:
      operationId: clearQueue
      tags: [queue]
      summary: Clear the queue
      responses:
        "204":
          description: Cleared
  /queue/{id}:
    delete:
      operationId: removeFromQueue
      tags: [queue]
      summary: Remove a queue item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
  /eq-presets:
    get:
      operationId: listEqPresets
      tags: [eq-presets]
      summary: List EQ presets
      responses:
        "200":
          description: Presets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EqPreset"
    post:
      operationId: createEqPreset
      tags: [eq-presets]
      summary: Create an EQ preset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEqPresetBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EqPreset"
  /eq-presets/active:
    put:
      operationId: setActiveEqPreset
      tags: [eq-presets]
      summary: Set the active EQ preset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetActiveEqPresetBody"
      responses:
        "200":
          description: Updated preset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EqPreset"
  /eq-presets/{id}:
    patch:
      operationId: updateEqPreset
      tags: [eq-presets]
      summary: Update an EQ preset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEqPresetBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EqPreset"
    delete:
      operationId: deleteEqPreset
      tags: [eq-presets]
      summary: Delete an EQ preset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Track:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        year:
          type: ["integer", "null"]
        genre:
          type: ["string", "null"]
        duration:
          type: number
        trackNumber:
          type: ["integer", "null"]
        fileName:
          type: string
        folderPath:
          type: string
        albumArtDataUrl:
          type: ["string", "null"]
        rating:
          type: integer
        source:
          type: string
        subsonicId:
          type: ["string", "null"]
        subsonicServerId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - artist
        - album
        - duration
        - fileName
        - folderPath
        - rating
        - source
        - createdAt
        - updatedAt
...
    Playlist:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        query:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - createdAt
        - updatedAt
    CreatePlaylistBody:
      type: object
      properties:
        name:
          type: string
        query:
          type: ["string", "null"]
      required:
        - name
    UpdatePlaylistBody:
      type: object
      properties:
        name:
          type: string
        query:
          type: ["string", "null"]
    PlaylistTrack:
      type: object
      properties:
        id:
          type: integer
        playlistId:
          type: integer
        trackId:
          type: integer
        position:
          type: integer
      required:
        - id
        - playlistId
        - trackId
        - position
    AddTrackToPlaylistBody:
      type: object
      properties:
        trackId:
          type: integer
        position:
          type: integer
      required:
        - trackId
    QueueItem:
      type: object
      properties:
        id:
          type: integer
        trackId:
          type: integer
        position:
          type: integer
        track:
          $ref: "#/components/schemas/Track"
      required:
        - id
        - trackId
        - position
        - track
    ReplaceQueueBody:
      type: object
      properties:
        trackIds:
          type: array
          items:
            type: integer
      required:
        - trackIds
    AddToQueueBody:
      type: object
      properties:
        trackIds:
          type: array
          items:
            type: integer
      required:
        - trackIds
    EqPreset:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        bands:
          type: string
        isActive:
          type: boolean
        isBuiltin:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - bands
        - isActive
        - isBuiltin
        - createdAt
        - updatedAt
    CreateEqPresetBody:
      type: object
      properties:
        name:
          type: string
        bands:
          type: string
        isActive:
          type: boolean
      required:
        - name
        - bands
    UpdateEqPresetBody:
      type: object
      properties:
        name:
          type: string
        bands:
          type: string
    SetActiveEqPresetBody:
      type: object
      properties:
        presetId:
          type: ["integer", "null"]
      required:
        - presetId
...