# lib/api-zod/src/generated/types/queueItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Audio Player API
 * OpenAPI spec version: 0.1.0
 */
import type { Track } from "./track";

export interface QueueItem {
  id: number;
  trackId: number;
  position: number;
  track: Track;
}