# lib/api-zod/src/generated/types/setActiveEqPresetBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Audio Player API
 * OpenAPI spec version: 0.1.0
 */

export interface SetActiveEqPresetBody {
  /** @nullable */
  presetId: number | null;
}