# lib/api-zod/src/generated/types/track.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Audio Player API
 * OpenAPI spec version: 0.1.0
 */

export interface Track {
  id: number;
  title: string;
  artist: string;
  album: string;
  /** @nullable */
  year?: number | null;
  /** @nullable */
  genre?: string | null;
  duration: number;
  /** @nullable */
  trackNumber?: number | null;
  fileName: string;
  folderPath: string;
  /** @nullable */
  albumArtDataUrl?: string | null;
  rating: number;
  source: string;
  /** @nullable */
  subsonicId?: string | null;
  /** @nullable */
  subsonicServerId?: number | null;
  createdAt: Date;
  updatedAt: Date;
}