# lib/api-zod/src/generated/types/updateEqPresetBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Audio Player API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateEqPresetBody {
  name?: string;
  bands?: string;
}